WIP just in case I blow all of this up.
the idea here is that I'm separating this out to a more protocol
oriented design. There's a base AX25Frame protocol that has basic stuff,
and then types that are domain specific to each frame type. It's
entirely possible I'll be making more either frame types, or kind of
a hierarchy of protocols, maybe. I'm not really sure yet.

But the main thing is I'm switching from a huge nasty initializer to
a Factory pattern and we'll see how it goes